Add a task to install the strings files back in plex-media-manager.

+++ Rakefile
@@ -2,6 +2,7 @@ require 'pathname'
 require File.expand_path('../lib/l10n', __FILE__)
 
 ROOT = Pathname(File.expand_path('..', __FILE__))
+PMM_ROOT = ROOT.parent+'plex-media-manager'
 
 def in_root(&block)
   Dir.chdir(ROOT, &block)
@@ -9,7 +10,7 @@ end
 
 include L10n
 
-file Localization['xx-chef'].strings => Localization['en'].strings do |f|
+file Localization['xx-chef'].strings => Localization['en'].strings do
   in_root do
     system %{bin/auto-translate xx-chef vendor/chef}
   end
@@ -25,11 +26,27 @@ end
 
 task :default => localizations.map {|localization| localization.xstrings }
 
-task :install => :default do
-  resources = ROOT.parent+'plex-media-manager/Resources'
+if PMM_ROOT.exist?
+  desc "Update the English strings from plex-media-manager"
+  task :update do
+    Dir.chdir(PMM_ROOT) do
+      puts "~ Updating English strings"
+      system %{genstrings -o Resources/en.lproj -s CPLocalizedString *.j */*.j}
 
-  localizations.each do |localization|
-    cp localization.strings, resources+"#{localization.code}.lproj/"
-    cp localization.xstrings, resources+"#{localization.code}.lproj/"
+      Dir.chdir(PMM_ROOT+'Resources/en.lproj') do
+        system %{iconv -f utf-16 -t utf-8 Localizable.strings > Localizable.strings.utf8 && mv Localizable.strings.utf8 Localizable.strings}
+        cp "Localizable.strings", Localization['en'].strings
+      end
+    end
+  end
+
+  desc "Install the .strings and .xstrings into plex-media-manager"
+  task :install => :default do
+    resources = PMM_ROOT+'Resources'
+
+    localizations.each do |localization|
+      cp localization.strings, resources+"#{localization.code}.lproj/"
+      cp localization.xstrings, resources+"#{localization.code}.lproj/"
+    end
   end
 end
